requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2
python-dotenv